% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualGrowthRateCalculator.R
\docType{methods}
\name{annualGrowthRateCalculator}
\alias{annualGrowthRateCalculator}
\title{Calculate annual growth rate}
\usage{
annualGrowthRateCalculator(
  boredDiameter,
  growthIncrement,
  growthYear,
  barkThickness
)
}
\arguments{
\item{boredDiameter}{numeric, Diameter at bored height in cm.}

\item{growthIncrement}{numeric, Growth increment in mm over a time period.}

\item{growthYear}{numeric, Number of years over which growth increment is measured.}

\item{barkThickness}{numeric, Bark thickness in mm. If missing, 0.05 will be used.}
}
\value{
Calculated annual growth rate.
}
\description{
This function is to calcualte annual growth rate.
}
\author{
Yong Luo
}
