% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEC2IC.R
\docType{methods}
\name{BEC2IC}
\alias{BEC2IC}
\alias{BEC2IC,character-method}
\title{Group BEC zones into interior and coastal region}
\usage{
BEC2IC(BEC)

\S4method{BEC2IC}{character}(BEC)
}
\arguments{
\item{BEC}{character, BC BEC zone(s)}
}
\value{
grouped region by bec zone, in which \code{C} stands for coastal region, \code{I} stands for interior region and
\code{?} stands for unknown region.
}
\description{
It groups the BC BEC zone into two regions: coastal region \code{C} and interior region \code{I}.
}
\author{
Yong Luo
}
