% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassCalculator.R
\docType{methods}
\name{biomassCalculator}
\alias{biomassCalculator}
\title{This function is to calculate aboveground biomass for boreal species based on DBH
or DBH/Height}
\usage{
biomassCalculator(
  species,
  DBH,
  heightIncluded = TRUE,
  height,
  paperSource = "Lambert2005"
)
}
\arguments{
\item{species}{Character string. The species name.}

\item{DBH}{Numeric. The tree's diameter at breast height (DBH, cm).}

\item{heightIncluded}{Logical. Whether the biomass is calculated based on DBH and height.
If TURE, height must be provided.
Default \code{TRUE}}

\item{height}{Numeric. The tree's height (m).}

\item{paperSource}{Character. Determine the sources of equations. Currently, this functions has two
options, i.e., "Lambert2005" and "Ung2008".
Default \code{Lambert2005}}
}
\value{
Biomass (kg) and missedSpecies list that was not calculated.
}
\description{
This function is to calculate aboveground biomass for boreal species based on DBH
or DBH/Height
}
\examples{
\dontrun{
 DBH <- seq(1, 100, 5)
 species <- c(rep("jack pine", 10), rep("black spruce", 10))
 species[1] <- "wrongSpecies"
 height <- seq(20, 40, length = 20)
 # without height information and taking the eqations from Lambert 2005
 biomass1 <- biomassCalculator(species = species, DBH = DBH, heightIncluded = FALSE)
 # with height information and taking the eqations from Lambert 2005
 biomass2 <- biomassCalculator(species = species, DBH = DBH,
                                heightIncluded = TRUE, height = height)
}
}
\author{
Yong Luo
}
