% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLD_remeas.R
\name{checkLD_remeas}
\alias{checkLD_remeas}
\title{Check the live and dead status of a remeasured subject}
\usage{
checkLD_remeas(subjectID, measNo, LDStatus, liveCode, deadCode)
}
\arguments{
\item{subjectID}{character, Specifies subject ID, such as a tree id.}

\item{measNo}{numeric, Measurement number with bigger value indicates a later measurement.}

\item{LDStatus}{character, Live and dead status for each remeasurement.}

\item{liveCode}{character, Code for live status.}

\item{deadCode}{character, Code for dead status.}
}
\value{
A data table that contains pass information. TRUE indicates pass, while FALSE indicates
failure.
}
\description{
This function is to check the live and deas status for a remeasured subject.
}
\author{
Yong Luo
}
