% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissing_remeas.R
\name{checkMissing_remeas}
\alias{checkMissing_remeas}
\title{Check the missing measurements in repeatedly measured subject against intended measurements}
\usage{
checkMissing_remeas(
  subjectID,
  measNo,
  intendedMeasNo,
  deadCode = NULL,
  LDStatus
)
}
\arguments{
\item{subjectID}{character, Specifies subject ID, such as a tree id.}

\item{measNo}{numeric, Measurement number with bigger value indicates a later measurement.}

\item{intendedMeasNo}{numeric, The measurement number that a subject is intended measured.}

\item{deadCode}{character, The code indicates the subject is dead. This arguement serves two purposes:
1) switch the function whether check a subject under dead scenario (i.e., the missing
measurement before last intended measurement). Setting \code{NULL} will turn off dead mode,
arguement \code{LDStatus} will not be used in the function.
2) if the dead mode is turn on (i.e., anything but \code{NULL}),
this term determines which code will be used for dead. In this case, arguement \code{LSStatus}
must be specified. By default, this term is set as \code{NULL}.}

\item{LDStatus}{character, Live or dead status. This arguement is called only when \code{deadCode}
is not set as \code{NULL}.}
}
\value{
A data table that contains pass information. TRUE indicates pass, while FALSE indicates
failure. The table also contains the missing measurements.
}
\description{
This function is to check missing measurements in repeatedly measured
subject against intended measurements. Note that this function allows the
regeneration, which means there may be missing measurements at the begining of
intended masurements. Additionally, the function may allow the missing measurements
after a subject died, depending on how \code{deadCode} is specified.
}
\author{
Yong Luo
}
