% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBHClassifier.R
\docType{methods}
\name{DBHClassifier}
\alias{DBHClassifier}
\title{Derive DBH class from DBH}
\usage{
DBHClassifier(DBH, classInterval = 5, maxDBH = 175)
}
\arguments{
\item{DBH}{numeric, Tree DBH.}

\item{classInterval}{numeric, The interval that used to categorize the DBH. If missing
5 cm is used.}

\item{maxDBH}{numeric, Upper class limit. DBH that surpasses this limit is groupped
in at this limit. If missing 175 is used.}
}
\value{
Classified DBH
}
\description{
This function derives DBH classes based on DBH.
This function is equivalent to dbh_cl.sas macro.
}
\author{
Yong Luo
}
