% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIB_ICalculator.R
\docType{methods}
\name{DIB_ICalculator}
\alias{DIB_ICalculator}
\alias{DIB_ICalculator,character,character,character,numeric,numeric,numeric,numeric-method}
\title{Calculate the inside-bark diameter at a given height}
\usage{
DIB_ICalculator(
  taperEquationForm,
  FIZorBEC,
  species,
  height_I,
  heightTotal,
  DBH,
  volMultiplier
)

\S4method{DIB_ICalculator}{character,character,character,numeric,numeric,numeric,numeric}(
  taperEquationForm,
  FIZorBEC,
  species,
  height_I,
  heightTotal,
  DBH,
  volMultiplier
)
}
\arguments{
\item{taperEquationForm}{character, Specifies a taper equation form one of KBEC, KBECQCI, KFIZ3.}

\item{FIZorBEC}{character, Specifies FIZ or BEC.}

\item{species}{character, Species code.}

\item{height_I}{numeric, Height from ground.}

\item{heightTotal}{numeric, Total height of a tree.}

\item{DBH}{numeric, Diameter at breast height.}

\item{volMultiplier}{numeric, Volume adjustment multiplier.}
}
\value{
Diameter inside bark
}
\description{
This function uses taper equation to calculate diameter inside bark at a given height.
It is equivalent to the subroutine of vol_tree_active_equation in vol_setup macro
}
\author{
Yong Luo
}
