% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightEstimateForBTOP.R
\docType{methods}
\name{heightEstimateForBTOP_D}
\alias{heightEstimateForBTOP_D}
\title{Estimate tree height for a broken top tree when DBH, inside bark diamater at broken top height, height at broken
are available}
\usage{
heightEstimateForBTOP_D(heightBTOP, DIBBTOP, DBH,
  taperEquationForm = "KBEC", FIZorBEC, species, volMultiplier = 1,
  SASOriginal = FALSE)
}
\arguments{
\item{heightBTOP}{numeric, Height of the broken top.}

\item{DIBBTOP}{numeric, Diameter inside bark at the height of the broken top.}

\item{DBH}{numeric, DBH of the tree, Must be given when BTOP is D.}

\item{taperEquationForm}{character, Specifies which taper equaiton will be used to estimate
tree height, currently supports KBEC, KBECQCI, KFIZ.
If missing, the function uses KBEC as default.}

\item{FIZorBEC}{character, Specifies which FIZ or BEC (depends on taperEquationForm) zones the tree located.}

\item{species}{character, Tree species.}

\item{volMultiplier}{numeric, Volume adjustment. If missing, 1 will be used.}

\item{SASOriginal}{logical, Specifies whether the original sas algrithm will be used for guess tree height
If missing, FALSE will be used.}
}
\value{
Total tree height
}
\description{
This is the second function to estimate a tree's height for a broken top tree.
A tree's height is esimated using height of the broken top (\code{heightBTOP}), inside bark diameter at broken height (\code{DIBBTOP})
and \code{DBH}. Specifically, this function guesses the tree height,
computes inside bark diameter at broken height (\code{heightBTOP}) using a taper equation,
compares it to an observed inside bark diameter and chooses the tree height that has closest value of inside bark diameter at broken.
For the broken top trees that have field projected height, total tree height also can be estimated using
\code{\link{heightEstimateForBTOP_H}}.
}
\seealso{
\code{\link{heightEstimateForBTOP_H}}
}
\author{
Yong Luo
}
