% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightEstimateForBTOP.R
\docType{methods}
\name{heightEstimateForBTOP_H}
\alias{heightEstimateForBTOP_H}
\title{Estimate tree height for a broken top tree when projected tree height is available}
\usage{
heightEstimateForBTOP_H(heightProjected)
}
\arguments{
\item{heightProjected}{numeric, Projected tree height in the field, must be non-NA value.}
}
\value{
Total tree height
}
\description{
This function is to esimate a broken top tree's height based on projected tree
height in the field (\code{heightProjected}). For the broken top trees that have diameter at broken
and broken top trees, total tree height also can be esimated using \code{\link{heightEstimateForBTOP_D}}.
}
\seealso{
\code{\link{heightEstimateForBTOP_D}}
}
\author{
Yong Luo
}
