% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_group.R
\docType{methods}
\name{lm_group}
\alias{lm_group}
\alias{lm_group,character,data.table,character-method}
\alias{lm_group,character,data.table,missing-method}
\title{Extended lm function by adding group functionality}
\usage{
lm_group(formula, data, groupBy, ...)

\S4method{lm_group}{character,data.table,character}(formula, data, groupBy, ...)

\S4method{lm_group}{character,data.table,missing}(formula, data, groupBy, ...)
}
\arguments{
\item{formula}{character, Linear model formula.}

\item{data}{data.table, The data used for the models.}

\item{groupBy}{character, Specifies variables that used for the group.}

\item{...}{see \code{\link{lm}} for the rest arguments.}
}
\value{
A list of regression analyses results
}
\description{
A generic function by adding grouping functionality in \code{\link{lm}} function.
}
\seealso{
\code{\link{lm}}
}
\author{
Yong Luo
}
