% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dupUpdate.R
\docType{methods}
\name{merge_dupUpdate}
\alias{merge_dupUpdate}
\alias{merge_dupUpdate,data.table,data.table,character,logical-method}
\alias{merge_dupUpdate,data.table,data.table,character,missing-method}
\title{Merge table and update values for duplicate column}
\usage{
merge_dupUpdate(x, y, by, updateDup, ...)

\S4method{merge_dupUpdate}{data.table,data.table,character,logical}(x, y, by, updateDup, ...)

\S4method{merge_dupUpdate}{data.table,data.table,character,missing}(x, y, by, updateDup, ...)
}
\arguments{
\item{x}{data.table, The first table for merging.}

\item{y}{data.table, The second table for merging.}

\item{by}{character, The key to merge two tables.}

\item{updateDup}{logical, Specifies whether update duplicate column in merged table when its information is
available in y table, which means update from the second table.
If missing, the function takes \code{TRUE}.}

\item{...}{see \code{\link{merge}} for rest of arguments.}
}
\value{
A merged table without duplicate columes. A warning message is given if the duplicate column
has different values.
}
\description{
This is an extended function for \code{\link{merge}} function by updating values for duplicate
column for the first, second or both tables.
}
\seealso{
\code{\link{merge}}
}
\author{
Yong Luo
}
