% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHFCalculator.R
\docType{methods}
\name{PHFCalculator}
\alias{PHFCalculator}
\title{Calculate tree per ha factor for both fix and variable area plot}
\usage{
PHFCalculator(sampleType, blowUp, treeWeight, plotWeight, treeBasalArea)
}
\arguments{
\item{sampleType}{character, Specifies how the plot is sampled among fixed area plot or variable area plot,
must be either V for variable area plot or F for fixed area plot.}

\item{blowUp}{numeric, Specifies the blowup factor. For fixed area plot, it is calculated as 1/plotarea.
For variable area plot, it is basal area factor (BAF).}

\item{treeWeight}{numeric, Specifies whether a tree is zero counted (tree is out),
one time counted (regular count) or two times counted (double counted)
in the walk through sampling protocal.}

\item{plotWeight}{numeric, Specifies how a plot is measured, i.e., full plot measured (valued as 1),
half plot measured (valued as 2) or quarter plot measured (valued as 4).}

\item{treeBasalArea}{numeric, When plot is measured using variable area plot, this value must be given,
otherwise, can be missing}
}
\value{
Tree per ha factor
}
\description{
Calculates tree per ha factor for both fix and variable area plots.
}
\author{
Yong Luo
}
