% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomStemMapping.R
\docType{methods}
\name{randomStemMapping}
\alias{randomStemMapping}
\title{The function is to generate stem mapping using random spatial distribution of
stratified size group for non-stem-mapped trees.}
\usage{
randomStemMapping(objectID, size, noofGroup = 5, plotSize, mapSize = 10000)
}
\arguments{
\item{objectID}{character, The unique object identifier. Must be unique.}

\item{size}{numeric, The size that used for compute competition index.}

\item{noofGroup}{numeric, Defines how many groups to distribute objects.
Default is 5.}

\item{plotSize}{numeric, The plot size for the objects.}

\item{mapSize}{numeric, The map size to distribute the objects.
Default is 10000 m2.}
}
\value{
a data table that has five columns, plotNumber, treeNumber, Year, IntraH and InterH
}
\description{
The function is to generate stem mapping using random spatial distribution of
stratified size group for non-stem-mapped trees.
}
\note{
no note
}
\seealso{
no
}
\author{
Yong Luo
}
