% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIInBC.R
\name{SIInBC}
\alias{SIInBC}
\title{Derive site index for a given spatial coverage or a spatial point}
\usage{
SIInBC(SIMapPath, spatialCoverage, species = "all", returnClass = "table")
}
\arguments{
\item{SIMapPath}{character, Specifies folder location of species index maps. Please request
all the maps from author and save them into your target folder.
The function only supports \code{TIFF} format. Currently those
maps were converted from BC Data catalogue.}

\item{spatialCoverage}{spatialPolygons or spatialPoints, Specifies spatial polygons or spatial points that need to
intersect.}

\item{species}{character, Must be one or some of 22 species.}

\item{returnClass}{character, Specifies the class you intended to return from either
\code{sp} or \code{table}. If missing, \code{table} will be used.}
}
\value{
the returned value depends on \code{returnClass} arguement and class of
\code{spatialCoverage} arguement. If \code{returnClass} is set as \code{table},
a table will be returned. If \code{returnClass} is set as \code{sp}, a raster layer
will be returned for SpatialPolygons* objects, while a SpatialPointDataframe will be
returned for SpatialPoints* objects.
}
\description{
This function is to derive species' site index for a given spatial coverage or
spatial points based on BC provincial species productivity maps.
}
\author{
Yong Luo
}
