% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesCode2speciesType.R
\docType{methods}
\name{speciesCode2speciesType}
\alias{speciesCode2speciesType}
\alias{speciesCode2speciesType,character-method}
\title{Group species into deciduous and coniferous species group}
\usage{
speciesCode2speciesType(species)

\S4method{speciesCode2speciesType}{character}(species)
}
\arguments{
\item{species}{character, Tree basic species code, which is SPECIES in VRI original data.}
}
\value{
Species type: \code{D}-deciduous species and \code{C}-coniferous species.
\code{NA}, with a warning message, is given if a species fails to be grouped.
}
\description{
This function is to group species into deciduous and
coniferous species types based on BC species code and \code{sp_type} lookup table.
This function uses hardcoded \code{sp_type} lookup table.
}
\author{
Yong Luo
}
