% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSpeciesName.R
\docType{methods}
\name{standardizeSpeciesName}
\alias{standardizeSpeciesName}
\alias{standardizeSpeciesName,data.table,character-method}
\title{Standardize species name from different forest inventory data, this function to make all
the species compatible to biomassCalculation function}
\usage{
standardizeSpeciesName(speciesTable, forestInventorySource)

\S4method{standardizeSpeciesName}{data.table,character}(speciesTable,
  forestInventorySource)
}
\arguments{
\item{speciesTable}{data table. It must at least have one column species}

\item{forestInventorySource, }{Character string. Give the forest inventory data source
Currently support MBPSP, MBTSP, ABPSP, BCPSP, SKPSP, SKTSP and NFIPSP}
}
\value{
a data tables, the first one contains successfully standardized species.
the newSpeciesName is the standardized name, unknown means the species in
the original species table can not be found according to manual
}
\description{
Standardize species name from different forest inventory data, this function to make all
the species compatible to biomassCalculation function
}
\note{
no note
}
\seealso{
no
}
\author{
Yong Luo
}
