% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemMapping.R
\docType{methods}
\name{stemMapping}
\alias{stemMapping}
\title{Map stems in a plot.}
\usage{
stemMapping(
  objectID,
  bearing,
  distance,
  plotShape = as.character(NA),
  plotSize,
  distanceUnit = "m",
  outputFormat,
  showID = TRUE
)
}
\arguments{
\item{objectID}{character, A object's ID, e.g., a tree's ID.}

\item{bearing}{numeric, Azimuth of a object from the north. It should be between 0 to 360.}

\item{distance}{numeric, Distance between a object and the centre of the circle.}

\item{plotShape}{character, circle, Rectangle or NA. NA means no plot shape information is not available and
and no plot geometry will be produced.}

\item{plotSize}{numeric, Defines the plot geoimetry. If plotShape is circular, a number must be provided as plot radius.
If plotShape is rectangle, two numbers must be provided to determine length and width of a
plot.}

\item{distanceUnit}{character, Defines the unit in the stem mapping. Default is \code{m}.}

\item{outputFormat}{character, Defines whether a \code{table} or a \code{figure} will be returned
from this function.}

\item{showID}{logical, Specifies whether the objectID will be showed in outputed figure. Default is \code{TRUE}.}
}
\value{
\enumerate{
\item A table contains the x and y for all the objects and the plot boundary.
2) a ggplot object will be returned for visualization.
}
}
\description{
The function is to map all the stems in a plot based on bearing and distance.
}
\note{
In the figure, IPC is the integrated plot center with x = 0 and y = 0.
}
\examples{
\dontrun{
## randomly generate some trees
bigplottrees <- data.table(tree_id = 1:20,
                           angle = runif(20, min = 0, max = 360),
                           distance = runif(20, min = 0, max = 11.28))
## output a table
stemmapped_table <- stemMapping(objectID = bigplottrees$tree_id,
                                         bearing = bigplottrees$angle,
                                         distance = bigplottrees$distance,
                                         plotShape = "circle",
                                         plotSize = 11.28,
                                         distanceUnit = "m",
                                         outputFormat = "table")

## output a figure
stemmapped_figure <- stemMapping(objectID = bigplottrees$tree_id,
                                         bearing = bigplottrees$angle,
                                         distance = bigplottrees$distance,
                                         plotShape = "circle",
                                         plotSize = 11.28,
                                         distanceUnit = "m",
                                         outputFormat = "figure",
                                         showID = TRUE)

## output a figure without a plot
stemmapped_figure <- stemMapping(objectID = bigplottrees$tree_id,
                                         bearing = bigplottrees$angle,
                                         distance = bigplottrees$distance,
                                         outputFormat = "figure",
                                         showID = TRUE)

}

}
\author{
Yong Luo
}
