% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIB_ICalculator.R
\docType{methods}
\name{taperCoeffsGenerator}
\alias{taperCoeffsGenerator}
\alias{taperCoeffsGenerator,character-method}
\alias{taperCoeffsGenerator,missing-method}
\title{Generate the coefficients table of taper equations}
\usage{
taperCoeffsGenerator(taperEquationForm)

\S4method{taperCoeffsGenerator}{character}(taperEquationForm)

\S4method{taperCoeffsGenerator}{missing}()
}
\arguments{
\item{taperEquationForm}{character, Specifies a taper equation form one of KBEC, KBECQCI, KFIZ3.}
}
\value{
A coeffients table
}
\description{
Generates the coefficients of the taper equations for based on specific
taper equation form (\code{taperEquationForm})
}
\author{
Yong Luo
}
