% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIB_ICalculator.R
\docType{methods}
\name{taperImplementor}
\alias{taperImplementor}
\alias{taperImplementor,character,data.table,character,character,numeric,numeric,numeric,numeric-method}
\title{Implement taper equation for a given tree}
\usage{
taperImplementor(
  taperEquationForm,
  taperCoeffs,
  FIZorBEC,
  species,
  height_I,
  heightTotal,
  DBH,
  volMultiplier
)

\S4method{taperImplementor}{character,data.table,character,character,numeric,numeric,numeric,numeric}(
  taperEquationForm,
  taperCoeffs,
  FIZorBEC,
  species,
  height_I,
  heightTotal,
  DBH,
  volMultiplier
)
}
\arguments{
\item{taperEquationForm}{character, Specifies a taper equation form one of KBEC, KBECQCI, KFIZ3.}

\item{taperCoeffs}{data.table, Table that stores the coefficients that match the taper equation.}

\item{FIZorBEC}{character, FIZ or BEC.}

\item{species}{character, Species code.}

\item{height_I}{numeric, Height from ground.}

\item{heightTotal}{numeric, Total height of a tree.}

\item{DBH}{numeric, Diameter at breast height.}

\item{volMultiplier, }{Volume multiplier adjustment.}
}
\value{
DIB_I diameter inside bark at height_I
}
\description{
Implement taper equation for a given tree
}
\note{
This function is inside of the VRIVolTree function
}
\author{
Yong Luo
}
