% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeProfile.R
\docType{methods}
\name{treeProfile}
\alias{treeProfile}
\title{Calculate volume for trees}
\usage{
treeProfile(
  taperEquationForm = "KBEC",
  FIZorBEC,
  species,
  height,
  DBH,
  stumpHeight = 0.3,
  breastHeight = 1.3,
  UTOPDIB = 10,
  BTOPHeight = NA
)
}
\arguments{
\item{taperEquationForm}{character, Specifies which taper equations will be used, currently support \code{KBEC} or \code{KFIZ3}.
\code{KBEC} is the Kozak's equations (2002 version) based on BEC zone, tree sizes and species.
\code{KFIZ3} is the equations based on forest inventory zone (FIZ), tree sizes and species.
Default is KBEC, if missing.}

\item{FIZorBEC}{character, Specifies which FIZ or BEC (depends on taperEquationForm) zones the tree located in BC.}

\item{species}{character, Tree species, must be BC species code.}

\item{height}{numeric, Total tree height in meter.}

\item{DBH}{numeric, DBH of the tree in cm.}

\item{stumpHeight}{numeric, Defines stump height. If missing, 0.3 m is used.}

\item{breastHeight}{numeric, Defines the breast height. If missing, 1.3 m is used.}

\item{UTOPDIB}{numeric, Merchantable inside-bark diameter. If missing, UTOP is 10.}

\item{BTOPHeight}{numeric, Height at broken top.}
}
\value{
A volume table
}
\description{
This function is to produce a tree trunk profile (i.e., inside bark diameter (\code{DIB})).
And summarize the whole stem volume (\code{VOL_WSV}) and merchantable volume (\code{VOL_MER}).
}
\note{
For the volume between 0 and 0.3, also known as stump volume,
the compiler calculates the volume as cylinder with the diameter of stump
height. In the case of the diameter at stump height is less than diameter at breast height,
the diameter at breast height is used as stump height. It calculates tree volume based on a 10 cm
slices starting from 0.3 m tall using Smalian’s formula.
}
\examples{
\dontrun{
treeprofile_a <- treeProfile(FIZorBEC = "CWH",
                             species = "H",
                             height = 27.4,
                             DBH = 30.7,
                             BTOPHeight = 5.6)
treeprofile_b <- treeProfile(FIZorBEC = "CWH",
                             species = "S",
                             height = 37.3,
                             DBH = 42.3)
treeprofile_c <- treeProfile(FIZorBEC = "CWH",
                             species = "H",
                             height = 11.6,
                             DBH = 11.2)

}

}
\seealso{
\code{\link{treeVolCalculator}}
}
\author{
Yong Luo
}
