% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeVolCalculator.R
\docType{methods}
\name{treeVolCalculator}
\alias{treeVolCalculator}
\title{Calculate volume for trees}
\usage{
treeVolCalculator(FIZorBEC, species, height, DBH,
  taperEquationForm = "KBEC", volMultiplier = 1, stumpHeight = 0.3,
  breastHeight = 1.3, UTOPDIB = 10, BTOPEstimateType = NA,
  BTOPHeight = NA, BTOPDIB = NA, logLengthMatrix = data.table(Log1_L
  = numeric()), logMinLength = 3)
}
\arguments{
\item{FIZorBEC}{character, Specifies which FIZ or BEC (depends on taperEquation) zones the tree located in BC.}

\item{species}{character, Tree species, must be BC species code.}

\item{height}{numeric, Total tree height in meter.}

\item{DBH}{numeric, DBH of the tree in cm.}

\item{taperEquationForm}{character, Specifies which taper equaiton will be used, currently support KFIZ3 or KBEC.
See function \code{DIB_ICalculator} for details. Default is KBEC, if missing.}

\item{volMultiplier}{numeric, Volume adjustment multiplier. If missing, 1 (no adjustment) is used.}

\item{stumpHeight}{numeric, Defines stump height. If missing, 0.3 m is used.}

\item{breastHeight}{numeric, Defines the breast height. If missing, 1.3 m is used.}

\item{UTOPDIB}{numeric, Merchantable inside-bark diameter. If missing, UTOP is 10.}

\item{BTOPEstimateType}{integer, Must among NA, 1, 2, 3. Defines whether a tree has broken top and which field observation (height at broken or DIB at broken )
is used to define broken point. NA means that tree is not broken top. 1 and 3 means diameter at broken top is not available,
height at broken top is used to define broken point. 2 means diameter at broken top is available and is used to define broken point.
Default is NA: tree does not have broken top.}

\item{BTOPHeight}{numeric, Height at broken top.}

\item{BTOPDIB}{numeric, Diameter inside bark at height of broken top.}

\item{logLengthMatrix}{data.table, Log length matrix. If missing, there is no log-level volume returned.}

\item{logMinLength}{numeric, Minimum log length. This argument is activated when logLengthMatrix is provided.}
}
\value{
A volume table
}
\description{
This function is to calculate tree volume using taper equations on a basis of 10 cm slice. As default, the function is to calculate whole tree volume
(\code{VOL_WSV}), total merchantable volume (\code{VOL_BELOW_UTOP}) and non-merchantable volume
(\code{VOL_ABOVE_UTOP}) based on \code{FIZorBEC}, \code{species}, \code{height}, \code{DBH} using Kozak BEC taper equations.
The function also handles broken top trees by specifying \code{BTOPEstimateType}, \code{BTOPHeight} and \code{BTOPDIB}.
Accordingly, \code{VOL_BELOW_BTOP} and \code{VOL_ABOVE_BTOP} are produced.
Lastly, the function derives volume (denoted as \code{LOG_V_X}), merchantable volume
(denoted as \code{LOG_VM_X}) and top inside bark diameter (denoted as \code{LOG_D_X}) for each log when the \code{logLengthMatrix}
is provided. For all the scenarioes, stump height (\code{HT_STUMP}), inside bark diameter at stump height (\code{DIB_STUMP}),
breast height (\code{HT_BH}), inside bark diameter at breast height (\code{DIB_BH}) are generated.
}
\author{
Yong Luo
}
