% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_Convertor.R
\docType{methods}
\name{UTM_Convertor}
\alias{UTM_Convertor}
\title{Convert UTM to other coordinate reference system.}
\usage{
UTM_Convertor(
  point_ID,
  zone,
  northing,
  easting,
 
    CRS_To = "+proj=aea +lat_0=45 +lon_0=-126 +lat_1=50 +lat_2=58.5 +x_0=1000000 +y_0=0 +datum=NAD83 +units=m +no_defs",
  class = "sp"
)
}
\arguments{
\item{point_ID}{character, Data point ID.}

\item{zone}{integer, UTM zone.}

\item{northing}{integer, UTM northing.}

\item{easting}{integer, UTM easting.}

\item{CRS_To}{character, Defines the spatial coordination reference that you wish to transform.
Default is BC Albers reference system.}

\item{class}{character, Define the class of returned objective. Currently this function supports
either \code{table} or \code{sp} class. Default is \code{table}.}
}
\value{
Reprojected objective.
}
\description{
Converts UTM coordinates to the other coordinate reference system.
}
\examples{
\dontrun{
## for Prince Rupert, Fort Nelson, Prince George, Victoria, Kelowna
citylocs <- UTM_Convertor(point_ID = c("Prince Rupert", "Prince George",
                                       "Victoria", "Kelowna"),
                          zone = c(9, 10, 10, 11),
                          northing = c(6019079.41, 5974323.27, 5361626.96, 5528467.98),
                          easting = c(415075.83, 516441.65, 475594.70, 321996.76),
                          class = "sp")
bcbdry <- bcmaps::bc_bound(class = "sp")
plot(bcbdry)
plot(citylocs, col = "red", size = 10, add = TRUE)
}

}
\author{
Yong Luo
}
