% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkThroughFlag_field.R
\docType{methods}
\name{walkThroughFlag_field}
\alias{walkThroughFlag_field}
\title{To determine a tree-level flag in a walk-through sample in the field}
\usage{
walkThroughFlag_field(
  bearingBoundary,
  distanceBoundary,
  radiusYSM = 11.28,
  treeList
)
}
\arguments{
\item{bearingBoundary}{numeric, The bearing (degree) for a walk-through edge.}

\item{distanceBoundary}{numeric, The distance (m) between IPC to the walk-through edge.}

\item{radiusYSM}{numeric, The radius (m) of a fixed area plot, default is 11.28, which is BC's YSM plot radius.}

\item{treeList}{list, Defines a tree number, bearing and distance of a tree from IPC.}
}
\value{
The function returns three objects: 1. a figure shows the plot and boundary and trees with wkthrough codes
2. a lookup table in which a field crew can identify a tree's wkthrough codes
based on a tree's bearing.
3. assigned wkthrough codes for the trees provided.
}
\description{
This function determines the walk-through area and assign walk-through flag for a tree
}
\note{
This method is adapted from Ducey 2004 paper. The population boundary is a
simple straight line, which is defined by the boundary bearing (\code{bearingBoundary})
and boundary distance (\code{distanceBoundary}).
Purple line (from P1 to P4) is the boundary. P1 to P4 are the four pins. The dark grey area
is the walkthrough area, in which all trees should be marked as \code{W}.
The white area in the circular plot is out of the plot, in which trees should be marked as \code{O}.
No mark should be made for trees in rest of the plot (light grey area).
pin1 (P1 in green) is always the right direction (90 degree on your right) from the distance point (D in red)
with your back to the IPC.
The bearing of boundary is clock-wise angle from the North to the D-P1. The distance
of boundary is between IPC and D.
}
\examples{
\dontrun{
# for bearing of 45 with distance of 4.5
plot_exam_fld <- walkThroughFlag_field(bearingBoundary = 45,
                              distanceBoundary = 11.30,
                              treeList = list(c(1, 30, 2),
                                              c(5, 120, 6),
                                              c(6, 160, 10),
                                              c(8, 250, 7),
                                              c(9, 110, 9),
                                              c(10, 340, 9),
                                              c(12, 340, 11),
                                              c(13, 320, 13)))
}
}
\author{
Yong Luo
}
