% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkThrough.R
\docType{methods}
\name{walkThroughPins}
\alias{walkThroughPins}
\title{To calculate 4 pins for a a walk-through sample}
\usage{
walkThroughPins(bearingBoundary, distanceBoundary, radiusYSM = 11.28)
}
\arguments{
\item{bearingBoundary}{numeric, The bearing (degree) for a walk-through edge.}

\item{distanceBoundary}{numeric, The distance (m) between IPC to the walk-through edge.}

\item{radiusYSM}{numeric, The radius (m) of a YSM plot, default is 11.28.}
}
\description{
This function is to calculate 4 pins for a walk-through sample
}
\note{
This function is equivalent to the excel provided by Chris, which is used in
handheld.
}
\author{
Yong Luo
}
