% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatial.R
\name{getSpatial}
\alias{getSpatial}
\title{Generic function to derive BEC, TSA and FIZ for given locations}
\usage{
getSpatial(pointID, zone, northing, easting, spatialAttribute = "all",
  mapPath = as.character(NA), mapName = as.character(NA),
  mapFormat = "gdb")
}
\arguments{
\item{pointID}{character, Data point ID.}

\item{zone}{integer, UTM zone.}

\item{northing}{integer, UTM northing.}

\item{easting}{integer, UTM easting.}

\item{spatialAttribute}{character, specifies which spatial attribute to be obtained. Must be one of
"BEC", "TSA", "FIZ", "TFL" and "OWNERSHIP", regardless of lower or upper cases.
The spatial attribute \code{BEC} and \code{TSA} are
available at \code{\link[bcmaps]{bec}} and \code{\link[bcmaps]{tsa}}.
Therefore, for these attributes, \code{mapPath}, \code{mapName} and
\code{mapFormat} should not be specified. Since \code{FIZ}, \code{TFL} and
\code{OWNERSHIP} map is not available from any R package.
The below arguements must be specified.}

\item{mapPath}{character, Path to map.}

\item{mapName}{character, Map name of the map.}

\item{mapFormat}{character, Map format of the map.}
}
\value{
Depends on what spatial attribute a function derives.
For \code{BEC}, a table that contains:
\itemize{
\item{bec_zone} {bec zone.}
\item{bec_sbz} {bec subzone.}
\item{bec_var} {bec variant}
}
For \code{TSA}, a table that contains:
\itemize{
\item{tsa} {tsa.}
\item{tsa_desc} {tsa descriptions.}
}
For \code{FIZ}, a table that contains:
\itemize{
\item{fiz} {fiz forest inventory zone.}
}
For \code{TFL}, a table that contains:
\itemize{
\item{tfl} {tfl timber farm licenses.}
}
For \code{OWNERSHIP}, a table that contains:
\itemize{
\item{owner} {owner of land.}
\item{schedule} {schedule.}
}
}
\description{
This function is to derive \code{BEC}, \code{TSA} or {FIZ} based on an UTM location and BEC map.
}
\author{
Yong Luo
}
