% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeProfile.R
\docType{methods}
\name{treeProfile}
\alias{treeProfile}
\title{Calculate volume for trees}
\usage{
treeProfile(
  taperEquationForm = "KBEC",
  FIZorBEC,
  species,
  height,
  DBH,
  stumpHeight = 0.3,
  breastHeight = 1.3,
  UTOPDIB = 10,
  BTOPHeight = NA
)
}
\arguments{
\item{taperEquationForm}{character, Specifies which taper equaiton will be used, currently support KFIZ3 or KBEC.
See function \code{DIB_ICalculator} for details. Default is KBEC, if missing.}

\item{FIZorBEC}{character, Specifies which FIZ or BEC (depends on taperEquation) zones the tree located in BC.}

\item{species}{character, Tree species, must be BC species code.}

\item{height}{numeric, Total tree height in meter.}

\item{DBH}{numeric, DBH of the tree in cm.}

\item{stumpHeight}{numeric, Defines stump height. If missing, 0.3 m is used.}

\item{breastHeight}{numeric, Defines the breast height. If missing, 1.3 m is used.}

\item{UTOPDIB}{numeric, Merchantable inside-bark diameter. If missing, UTOP is 10.}

\item{BTOPHeight}{numeric, Height at broken top.}
}
\value{
A volume table
}
\description{
This function is to calculate tree volume using taper equations on a basis of 10 cm slice. As default, the function is to calculate whole tree volume
(\code{VOL_WSV}), total merchantable volume (\code{VOL_BELOW_UTOP}) and non-merchantable volume
(\code{VOL_ABOVE_UTOP}) based on \code{FIZorBEC}, \code{species}, \code{height}, \code{DBH} using Kozak BEC taper equations.
The function also handles broken top trees by specifying \code{BTOPEstimateType}, \code{BTOPHeight} and \code{BTOPDIB}.
Accordingly, \code{VOL_BELOW_BTOP} and \code{VOL_ABOVE_BTOP} are produced.
Lastly, the function derives volume (denoted as \code{LOG_V_X}), merchantable volume
(denoted as \code{LOG_VM_X}) and top inside bark diameter (denoted as \code{LOG_D_X}) for each log when the \code{logLengthMatrix}
is provided. For all the scenarioes, stump height (\code{HT_STUMP}), inside bark diameter at stump height (\code{DIB_STUMP}),
breast height (\code{HT_BH}), inside bark diameter at breast height (\code{DIB_BH}) are generated.
}
\examples{
\dontrun{
treeprofile_a <- treeProfile(FIZ = "CWH",
                             species = "H",
                             height = 27.4,
                             DBH = 30.7,
                             BTOPHeight = 5.6)
treeprofile_b <- treeProfile(FIZ = "CWH",
                             species = "S",
                             height = 37.3,
                             DBH = 42.3)
treeprofile_c <- treeProfile(FIZ = "CWH",
                             species = "H",
                             height = 11.6,
                             DBH = 11.2)

}

}
\author{
Yong Luo
}
