% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeVolume.R
\docType{methods}
\name{treeVolume}
\alias{treeVolume}
\title{Calculate volume for trees}
\usage{
treeVolume(
  taperEquationForm = "KBEC",
  FIZorBEC,
  species,
  DBH,
  height,
  BTOPHeight = NA,
  volumeName = "WSV",
  stumpHeight = 0.3,
  UTOPDIB = 10
)
}
\arguments{
\item{taperEquationForm}{character, Specifies which taper equaiton will be used, currently support KFIZ3 or KBEC.
See function \code{\link{DIB_ICalculator}} for details. Default is KBEC, if missing.}

\item{FIZorBEC}{character, Specifies which FIZ or BEC (depends on taperEquation) zones the tree located in BC.}

\item{species}{character, Tree species, must be BC species code.}

\item{DBH}{numeric, DBH of the tree in cm.}

\item{height}{numeric, Total tree height in meter.}

\item{BTOPHeight}{numeric, Height at broken top in meter. \code{NA} suggests no broken top.}

\item{volumeName}{character, Indicates which volume you want to derive. It supports the
whole stem volume (\code{WSV}), merchantable volume (\code{MER})
or stump volume (\code{STUMP}).
The merchantable volume is the whole stem volume minus volume of
stump and volume less than minimum utility diameter (\code{UTOPDIB}).}

\item{stumpHeight}{numeric, Defines stump height. If missing, 0.3 m is used. It will be called to
calculate the merchantable volume.}

\item{UTOPDIB}{numeric, Merchantable inside-bark diameter. If missing, UTOP is 10. It will be called to
calculate the merchantable volume.}
}
\value{
volume
}
\description{
This function is to produce a tree trunk profile (i.e., inside bark diameter (\code{DIB})).
And summarize the whole stem volume (\code{VOL_WSV}) and merchantable volume (\code{VOL_MER}).
}
\note{
For the volume between 0 and 0.3, also known as stump volume,
the function calculates the volume as cylinder with the diameter of stump
height. In the case of the diameter at stump height is less than diameter at breast height,
the diameter at breast height is used as stump height. It calculates tree volume based on a 10 cm
slices starting from 0.3 m tall using Smalian’s formula.
}
\examples{
\dontrun{
treeA_wsv <- treeVolume(FIZorBEC = "CWH",
                        species = "H",
                        DBH = 30.7,
                        height = 27.4,
                        BTOPHeight = 5.6)
trees_wsv <- treeVolume(FIZorBEC = "CWH",
                        species = c("H", "S", "H"),
                        DBH = c(30.7, 42.3, 11.2),
                        height = c(27.4, 37.3, 11.6),
                        BTOPHeight = c(5.6, NA, NA))
trees_mer <- treeVolume(FIZorBEC = "CWH",
                        volumeName = "MER",
                        species = c("H", "S", "H"),
                        DBH = c(30.7, 42.3, 11.2),
                        height = c(27.4, 37.3, 11.6),
                        BTOPHeight = c(5.6, NA, NA))
}

}
\seealso{
\code{\link{treeVolCalculator}} and \code{\link{treeProfile}}
}
\author{
Yong Luo
}
